move expression factory manager into a new file
